+++ src/Either/Right.php
@@ -92,4 +92,10 @@ final class Right implements Implementation
     {
         return new Left($this->value);
     }
+
+    public function eitherWay(callable $right, callable $left): Either
+    {
+        /** @psalm-suppress ImpureFunctionCall */
+        return $right($this->value);
+    }
 }