+++ tests/EitherTest.php
@@ -490,4 +490,79 @@ class EitherTest extends TestCase
                 );
             });
     }
+
+    public function testEitherWay()
+    {
+        $this
+            ->forAll(
+                Set\AnyType::any(),
+                Set\AnyType::any(),
+            )
+            ->then(function($initial, $new) {
+                $this->assertSame(
+                    $new,
+                    Either::left($initial)
+                        ->eitherWay(
+                            static fn() => Either::left($initial),
+                            function($value) use ($initial, $new) {
+                                $this->assertSame($initial, $value);
+
+                                return Either::right($new);
+                            },
+                        )
+                        ->match(
+                            static fn($value) => $value,
+                            static fn() => null,
+                        ),
+                );
+                $this->assertSame(
+                    $new,
+                    Either::defer(static fn() => Either::left($initial))
+                        ->eitherWay(
+                            static fn() => Either::left($initial),
+                            function($value) use ($initial, $new) {
+                                $this->assertSame($initial, $value);
+
+                                return Either::right($new);
+                            },
+                        )
+                        ->match(
+                            static fn($value) => $value,
+                            static fn() => null,
+                        ),
+                );
+                $this->assertSame(
+                    $new,
+                    Either::right($initial)
+                        ->eitherWay(
+                            function($value) use ($initial, $new) {
+                                $this->assertSame($initial, $value);
+
+                                return Either::right($new);
+                            },
+                            static fn() => Either::left($initial),
+                        )
+                        ->match(
+                            static fn($value) => $value,
+                            static fn() => null,
+                        ),
+                );
+                $this->assertSame(
+                    $new,
+                    Either::defer(static fn() => Either::right($initial))
+                        ->eitherWay(
+                            function($value) use ($initial, $new) {
+                                $this->assertSame($initial, $value);
+
+                                return Either::right($new);
+                            },
+                            static fn() => Either::left($initial),
+                        )
+                        ->match(
+                            static fn($value) => $value,
+                            static fn() => null,
+                        ),
+                );
+            });
+    }
 }